requirements.txt
--find-links https://download.pytorch.org/whl/cpu
flask==3.0.0
torch==2.8.0
torchvision==0.23.0
--find-links https://download.pytorch.org/whl/torchaudio
torchaudio==2.8.0
Pillow>=10.2.0
transformers==4.36.2
sentence-transformers==2.2.2
huggingface-hub>=0.20.3
gunicorn==21.2.0
werkzeug==3.0.0
numpy>=1.22.0
tqdm>=4.65.0